Module lcd_example::lib::lcd_utils

Constants

ASCII_OFFSET 🔒

Provided in order to convert u8 into proper ASCII char

BIT_0 🔒

BIT_0 through BIT_7 are bit masks in order to process the ASCII char parameter

BIT_1 🔒

BIT_2 🔒

BIT_3 🔒

BIT_4 🔒

BIT_5 🔒

BIT_6 🔒

BIT_7 🔒

D0_PIN 🔒

D1_PIN 🔒

D2_PIN 🔒

D3_PIN 🔒

D4_PIN 🔒

D5_PIN 🔒

D6_PIN 🔒

D7_PIN 🔒

ENABLE_PIN 🔒

LED_PIN 🔒

All pin macro numbers correspond to their respective index the in the gpio array in main()

MAX_LENGTH 🔒

Maximum length for one row of characters

MS_DELAY 🔒

Arbitrary value, user-defined

READ_WRITE_PIN 🔒

REGISTER_SEL 🔒

Functions

clear_data_bits

Clears all data pins (D0-D7)

float_to_array

Function converts a f32 into a [char] array in order to be processed by the LCD display

reset_lcd

Clears the LCD display, and sets the cursor back to the beginning

test_write_char

Fills whole LCD display with characters as a test

write_char_array

Writes a [char] array to the LCD display

write_static_string

Writes a static &str to the LCD display